requirements.txt
selenium
pytest
pytest-html
google-generativeai
python-dotenv
